Improve message_id generation speed.
Messages ids are currently UUID4, and can't be yet replaced by a faster
and more efficient format because of some backward compatibilities
issues.

The solution here is to generate an UUID4 as a base, and then replace
its "time_low" part (4 bytes) by a randomly chosen unsigned integer
(32bits) incrementing until a randomly chosen limit.  When the limit is
reached, a new UUID4 is generated and a new value is randomly chosen for
the unsigned integer and its limit.

Speedup is between ~4x and ~20x depending of the speed of the local
entropy generation.