Enhance concurrent load on one zmq socket
Handle better the ZMQ/gevent bug that lead to miss change of state
events on a zmq socket. Also take care of the gevent bug of timeout that
triggers too early.

The initial workaround for the missing event bug was subject to the
gevent timeout early triggering. The workaround of one bug hitting a
second bug in its resolution! Yay! So much fun!

I tested everything with a simple echo service, a relay service which,
as its name says, relay messages to the echo service, and two client
connected to the relay service.

With one client connected, everything was fine, but with more than one,
the relay service was slowing down (by 90%!!!) because of the
combination of the missing event bug and the early triggering of timeout
one.

This patch reduce a little the throughput with one client (8.3% of loss)
because it involve more polling. In exchange adding more client doesn't
kill the performances anymore. And the relay and echo services are able
to tackle more requests per seconds as you add more clients (this is a
good sign).

Hopefully this wont introduce any other magical combination of any other
weird an obscure bug between ZMQ and Gevent...