update test, to help reproduce bug in travis
It seems that Travis has something special that makes some test crash
pretty bad. Its really hard if not impossible to reproduce, for now I
extended the heartbeat frequency of the tests to 2 seconds (to be longer
than the maximum wait on the (py)ZMQ bug of missing events). I also used
a gevent.Pool to forward exceptions.