Allow exporting non-hashable function
Instead of hashing the function object, hash its name instead. There is
no particular reason to consider a function instance and its name as two
independent entities in the context of the same object.

As a bonus, this means you can now replace your methods on the fly and
the next call will reach the new one. I am not sure why you would want
to do that though.