Python 2.6 compat: get the PyZmq Frame Buffer
A pyzmq frame in python2.6 and 2.7 are not similar. They both offer the
attribute buffer which is:
  - a read-only buffer in 2.6
  - a memoryview in 2.7

Apparently the read-only buffer doesn't implement the buffer interface
while the memoryview does, which is needed for msgpack.

Both accept to be sliced though which returns:
  - an 'str' in 2.6
  - a memoryview in 2.7

Sadly, until very recently, taking a slice of a memoryview of dimension
raises an exception: https://bugs.python.org/issue15944.
Although it looks like recent version of pyzqm>14.2 are not affected.

Anyway, the easiest is access the buffer differently depending of the
python version. Fixes 0rpc/zerorpc-python#125.

Additionally fixes the unit tests for 2.6 and add it back to travis.