Conversion to >=Python3.4
Code is now using the python 3 syntax , but thanks to "future" and
"builtins", it is compatible with 2.6 and 2.7.

Hopefully everything was done carefully enough to not impair
performances on python 2.

Note: Code is not functional and tests are not passing yet. A throughout
review of str, bytes and unicode strings is still necessary.