Finally correctly handling msgpack bin vs string.
A long time ago msgpack had a single RAW type for bytes string. It was
the application responsibility to use any compatible string encoding
across languages/runtimes.

At some point msgpack was updated with a distinction between BIN type
and STRING type. The STRING type is the RAW type renamed, and should
only contain and UTF-8 encoded string. The BIN type is to contain any
byte strings encoded at the application's responsibility. Because
STRING is the RAW type renamed, when reading a msgpack message written
an older version of msgpack, it is impossible to know if the string was
supposed to be encoded in UTF-8 or not.

So msgpack-python for backward compatibility, reads both RAW and BIN by
as bytes strings by default. On the write side bytes strings and unicode
strings encoded to UTF-8 are sent as STRING (and it will appear as UTF-8
string on nodejs for example).

And btw Python2 strings are somewhat loosely defined ASCII strings
(which is a subset of UTF-8) so everything was all fine, in/out of
zerorpc-python was UTF-8 strings.

But one day, somebody wanted zerorpc-python to handle the distinction
between unicode strings and bytes strings. This was the day commit
6ca6fdf598905ca67f2678973034dae435a28594 was born.

This broke the compatibility with zerorpc-node because suddenly,
zerorpc-python on python2 (remember, strings on python2 are bytes
strings) was sending BIN (bytes strings) instead of STRING (utf-8
strings) that zerorpc-node expects.

On the other hand, zerorpc-python on python3 was sending STRING as it
should, because in python3, strings are unicode by default and
msgpack-python will encode them as UTF-8 strings with the type STRING.

In an attempt of making zerorpc-python working across python2 &
python3, I did the wrong thing and used Python2 as reference instead of
using Python3, which further broke compatibility with zerorpc-node.

This commit restores compatibility with zerorpc-node and across
python3/python3.

Fixes #142.
Updates #108.

Note also this new repo 0rpc/zerorpc-crosstests to run basic zerorpc
tests across languages. I hope it will prevent this kind of bugs to be
published ever again.